Change Utf8Iterator::_width from int to size_t
There's no real issue with using an int, but it makes the intent
clearer and silences an MSVC warning.

+++ nokogumbo-import/gumbo-parser/src/utf8.c
@@ -106,7 +106,7 @@ static void add_error(Utf8Iterator* iter, GumboErrorType type) {
   // yet (and can't be, because it's invalid), so we need to build up the raw
   // hex value from the bytes under the cursor.
   uint32_t code_point = 0;
-  for (int i = 0; i < iter->_width; ++i) {
+  for (size_t i = 0; i < iter->_width; ++i) {
     code_point = (code_point << 8) | (unsigned char) iter->_start[i];
   }
   error->v.codepoint = code_point;
